changed variant id to big int

+++ database/migrations/2015_04_25_400000_create_variants_table.php
@@ -14,7 +14,7 @@ class CreateVariantsTable extends Migration {
     {
         Schema::create('variants', function(Blueprint $table)
         {
-            $table->integer('id')->unique();
+            $table->bigInteger('id')->unique();
             $table->integer('product_id');
             $table->integer('inventory_limit');
             $table->boolean('track')->default(True);
